Semana5

+++ views/DonacionesView.php
@@ -54,12 +54,15 @@ class DonacionesView
                 </dl>
                 <div class="mt-6 flex gap-2">
                     <a href="index.php?controller=donaciones&action=list" class="px-4 py-2 bg-blue-600 text-white rounded hover:bg-blue-700">Volver a la lista</a>
-                    <a href="index.php?controller=donaciones&action=edit&id=<?= urlencode($donacion['id'] ?? '') ?>" class="px-4 py-2 bg-yellow-500 text-white rounded hover:bg-yellow-600">
-                        <i class="fa fa-pencil" aria-hidden="true"></i> Editar
-                    </a>
-                    <a href="index.php?controller=donaciones&action=delete&id=<?= urlencode($donacion['id'] ?? '') ?>" class="px-4 py-2 bg-red-600 text-white rounded hover:bg-red-700" onclick="return confirm('¿Seguro que deseas eliminar esta donación?');">
-                        <i class="fa fa-trash" aria-hidden="true"></i> Eliminar
-                    </a>
+
+                    <?php if (!empty($_SESSION['rol']) && $_SESSION['rol'] === 'admin'): ?>
+                        <a href="index.php?controller=donaciones&action=edit&id=<?= urlencode($donacion['id'] ?? '') ?>" class="px-4 py-2 bg-yellow-500 text-white rounded hover:bg-yellow-600">
+                            <i class="fa fa-pencil" aria-hidden="true"></i> Editar
+                        </a>
+                        <a href="index.php?controller=donaciones&action=delete&id=<?= urlencode($donacion['id'] ?? '') ?>" class="px-4 py-2 bg-red-600 text-white rounded hover:bg-red-700" onclick="return confirm('¿Seguro que deseas eliminar esta donación?');">
+                            <i class="fa fa-trash" aria-hidden="true"></i> Eliminar
+                        </a>
+                    <?php endif; ?>
                 </div>
             </div>
         <?php endif; ?>
